requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
Faker==19.6.2
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
psycopg2==2.9.7
pycodestyle==2.11.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
